requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pandas==2.2.1
numpy==1.26.4
requests==2.31.0
pydantic==2.6.4
python-multipart==0.0.9
websockets==12.0
dhanhq
python-dotenv==1.0.0
yfinance==0.2.36